ncdas.cc: Modified the way attributes with type NC_BYTE are represented. Now the values are cast to 'int' before the string representation is made. This prevents values above 128 from making odd print reps which show up as blanks or unicode caharacters.

+++ ncdas.cc
@@ -100,7 +100,7 @@ print_attr(nc_type type, int loc, void *vals)
         gp.cp = (char *) vals;
 
         uc = *(gp.cp+loc);
-        rep << uc;
+        rep << (int)uc;
         return rep.str();
 
       case NC_CHAR:
